popup comment section with appropriate styling
