+++ src/style.css
@@ -6,6 +6,7 @@
   margin-left: 0;
   margin-right: 0;
   margin-top: 0;
+  z-index: 0;
 }
 
 html {
